--- a/build.boot
+++ b/build.boot
@@ -13,7 +13,6 @@
                     ; alda.core
                     [org.clojure/clojure    "1.8.0"]
                     [org.clojure/core.async "0.2.395"]
-                    [instaparse             "1.4.3"]
                     [com.taoensso/timbre    "4.7.4"]
                     [djy                    "0.1.4"]
                     [potemkin               "0.4.3"]