--- a/examples/dot_accessor.alda
+++ b/examples/dot_accessor.alda
@@ -1,0 +1,2 @@
+violin/viola/cello "strings": g1~1~1
+strings.cello: < c1~1~1