--- a/grammar/clojure-cached.bnf
+++ b/grammar/clojure-cached.bnf
@@ -1,1 +1,0 @@
-clj-expr-cached = <"⚙"> #"clj-expr\d+" <"⚙">