--- a/grammar/cram.bnf
+++ b/grammar/cram.bnf
@@ -1,2 +1,0 @@
-cram          = <"{"> <ows> events <ows> <"}"> duration?
-