--- a/grammar/event-sequence.bnf
+++ b/grammar/event-sequence.bnf
@@ -1,2 +1,0 @@
-event-sequence          = <"["> <ows> events? <ows> <"]">
-