--- a/grammar/header.bnf
+++ b/grammar/header.bnf
@@ -1,1 +1,0 @@
-header = (clj-expr-cached / set-variable / <ows>)*