--- a/grammar/numbers.bnf
+++ b/grammar/numbers.bnf
@@ -1,3 +1,0 @@
-<number>                = positive-number | negative-number
-positive-number         = #"[0-9]+"
-negative-number         = #"-[0-9]+"