--- a/grammar/score.bnf
+++ b/grammar/score.bnf
@@ -1,9 +1,0 @@
-score        = <ows> header? (<ows> part)*
-header       = !calls <ows> music-data+
-part         = calls <ows> music-data*
-calls        = name
-               (<ows> <"/"> <ows> name)*
-               (<ows> nickname)?
-               <ows> <":">
-<music-data> = !calls #"[a-zA-Z]*[^a-zA-Z]*([a-zA-Z][^a-zA-Z]+)*"
-