--- a/grammar/variables.bnf
+++ b/grammar/variables.bnf
@@ -1,6 +1,0 @@
-<variable-name> = #"[a-zA-Z]{2}[\w_]*"
-
-get-variable    = variable-name
-
-set-variable    = variable-name <ows-inline> <"="> <ows-inline>
-                  (events-inline / event-sequence)