--- a/grammar/voices.bnf
+++ b/grammar/voices.bnf
@@ -1,5 +1,0 @@
-<voices>     = voice (<ows> voice)* (<ows> (voice-zero | <#"\z"> | &<"]">))
-voice        = voice-number <ows> events-inside-voice
-voice-number = <"V"> #"[1-9]\d*" <":">
-voice-zero = <"V0:">
-