--- a/src/alda/lisp/code.clj
+++ b/src/alda/lisp/code.clj
@@ -1,12 +1,9 @@
 (ns alda.lisp.code
-  (:require [alda.parser-util :refer (parse-to-events-with-context)]))
+  (:require [alda.parser :refer (parse-input)]))
 
 (defn alda-code
   "Attempts to parse a string of text within the context of the current score;
    if the code parses successfully, the result is one or more events that are
    spliced into the score."
   [code]
-  (let [[context parse-result] (parse-to-events-with-context code)]
-    (if (= context :parse-failure)
-      (throw (Exception. (str "Invalid Alda code: " code)))
-      parse-result)))
+  (parse-input code :output :events))