--- a/src/alda/parser/parse_events.clj
+++ b/src/alda/parser/parse_events.clj
@@ -518,7 +518,7 @@
 (defn start-parsing-chord
   [parser token]
   (when (and (token-is :slash token)
-             (#{:note :rest :duration} (current-event-type parser)))
+             (not= :instrument-call (last-open-event parser)))
     (-> parser (assoc :chord? true))))
 
 (defn continue-parsing-chord