This is a rewrite of the Alda parser from the ground-up, replacing the current parser-generator (Instaparse)  implementation with an asynchronous pipeline using core.async. 

The motivation for the rewrite is described in detail in #20. Parsing is currently the bottleneck when parsing large scores -- it can often take seconds, depending on the size of the score. This rewrite speeds up parsing significantly for larger scores, while remaining about the same (just slightly faster) for scores under about 100 lines, like most of our example scores -- see the benchmarks below.

The idea is that we can start to construct a score in another thread as soon as the first token is parsed. There are actually several threads going, each performing a separate stage of the parsing process in parallel. The result is much faster parsing times (depending on the size of the score) and a foundation for doing more interesting things with streaming in the future.

### Benchmarks (parsing the example scores):

The metric here is the time it takes to parse an input file as events and then use the events to build a score. The "before" numbers are on the master branch, and the "after" numbers are on this branch.

```
across_the_sea.alda
  before: 42 ms
  after: 43 ms

awobmolg.alda
  before: 296 ms
  after: 180 ms

bach_cello_suite_no_1.alda
  before: 593 ms
  after: 368 ms

clapping_music.alda
  before: 186 ms
  after: 191 ms

debussy_quartet.alda
  before: 316 ms
  after: 267 ms

entropy.alda
  before: 105 ms
  after: 87 ms

gau.alda
  before: 66 ms
  after: 90 ms

hello_world.alda
  before: 4 ms
  after: 9 ms

key_signature.alda
  before: 37 ms
  after: 21 ms

multi-poly.alda
  before: 21 ms
  after: 12 ms

nesting.alda
  before: 29 ms
  after: 34 ms

overriding-a-global-attribute.alda
  before: 31 ms
  after: 20 ms

panning.alda
  before: 42 ms
  after: 49 ms

percussion.alda
  before: 42 ms
  after: 134 ms

phase.alda
  before: 34 ms
  after: 42 ms

poly.alda
  before: 291 ms
  after: 192 ms

printing.alda
  before: 21 ms
  after: 34 ms

variables.alda
  before: 96 ms
  after: 78 ms

variables-2.alda
  before: 12 ms
  after: 21 ms
```

As a final benchmark to test a very long score, I edited `bach_cello_suite_no_1.alda` by copying the entire file and pasting it to the end of the file about 15 times, the result being a score that is 15x the length of that example (just under 1000 lines of code):

```
Before: 8746 ms
After: 4709 ms
```

Note that this is only implements the "parse -> eval" piece of a fully streaming parse/play pipeline. I found that just by rewriting the parser by hand instead of using a parser generator library, parsing is significantly faster, but we still must wait for the score to be fully evaluated before we start to play it.

There is room for more improvement in the future -- we could start to play notes as soon as they are generated during the score-building phase. In the case of the 1000 line example above, this would take the 4709 ms wait before you start to hear any notes down to almost 0 ms! The score would still take almost 5 seconds to fully parse, but that would be happening in the background as you're listening to the score. I'm really excited about that potential, and I think this PR paves the way for us to get there in the near future.