@daveyarwood 2017-05-10
An added benefit of rewriting the parser from the ground up is that it fixes issue #12. Line and column numbers are propagated through correctly this way, and error messages are clearer and more precise.