@daveyarwood 2017-05-12
Trying this locally. Unfortunately, I'm hearing a lot of issues where audio playback will intermittently hang onto a note for a couple seconds or so before moving on. All the notes are in sync, it's just like the score pauses for a second before playback continues.

It seems like this is a weird interaction between JSyn and core.async.