@daveyarwood 2017-05-27
Good news! The problem only seems to happen when running in development mode via Boot. I think Boot imposes a good deal of overhead, so I think adding core.async to the mix might have been the straw that broke the camel's back.

I did a build of Alda using this branch, though, and ran it as an executable jar file, and the problem doesn't exist. Also, parsing is crazy fast. Around 100ms to parse the Bach cello suite example!

I feel good about including this in 1.0.0.